Sketch of localfilesystem and commiting

+++ fsspec/spec.py
@@ -1,4 +1,4 @@
-
+from contextlib import contextmanager
 from .utils import read_block
 
 
@@ -16,6 +16,9 @@ class AbstractFileSystem(object):
 
         A reasonable default should be provided if there are no arguments
         """
+        self.autocommit = True
+        self.files = None
+        self._intrans = False
         self._singleton[0] = self
 
     @classmethod
@@ -41,6 +44,16 @@ class AbstractFileSystem(object):
         """
         pass
 
+    @contextmanager
+    def transaction(self):
+        self.files = []
+        self._intrans = True
+        yield
+        for f in self.files:
+            f.commit()
+        self.files = None
+        self._intrans = False
+
     def mkdir(self, path, **kwargs):
         """
         Create directory entry at path
@@ -56,6 +69,23 @@ class AbstractFileSystem(object):
             may be permissions, etc.
         """
         pass
+
+    makedir = mkdir
+
+    def makedirs(self, path, exist_ok=False):
+        """Recursively make directories
+
+        Creates directory at path and any intervening required directories.
+        Raises exception if, for instance, the path already exists but is a
+        file.
+
+        Parameters
+        ----------
+        path: str
+            leaf directory name
+        exist_ok: bool (False)
+            If True, will error if the target already exists
+        """
 
     def rmdir(self, path):
         """Remove a directory, if empty"""
@@ -87,14 +117,24 @@ class AbstractFileSystem(object):
         """
         pass
 
-    def walk(self, path, detail=False):
+    listdir = ls
+
+    def walk(self, path, simple=False):
         """ Return all files belows path
 
-        Like ``ls``, but recursing into subdirectories. If detail is False,
-        returns a list of full paths.
+        Similar to ``ls``, but recursing into subdirectories.
+
+        Parameters
+        ----------
+        path: str
+            Root to recurse into
+        simple: bool (False)
+            If True, returns a list of filenames. If False, returns an
+            iterator over tuples like
+            (dirpath, dirnames, filenames), see ``os.walk``.
         """
 
-    def du(self, path, total=False, deep=False):
+    def du(self, path, total=False):
         """Space used by files within a path
 
         If total is True, returns a number (bytes), if False, returns a
@@ -103,14 +143,12 @@ class AbstractFileSystem(object):
         Parameters
         ----------
         total: bool
-            whether to sum all the file sized
-        deep: bool
-            whether to descend into subdirectories.
+            whether to sum all the file sizes
         """
-        if deep:
-            sizes = {f['name']: f['size'] for f in self.walk(path, True)}
-        else:
-            sizes = {f['name']: f['size'] for f in self.ls(path, True)}
+        sizes = {}
+        for f in self.walk(path, True):
+            info = self.info(f)
+            sizes[info['name']] = info['size']
         if total:
             return sum(sizes.values())
         else:
@@ -147,20 +185,35 @@ class AbstractFileSystem(object):
 
     def exists(self, path):
         """Is there a file at the given path"""
-        pass
+        try:
+            self.info(path)
+            return True
+        except:
+            return False
 
     def info(self, path):
         """Give details of entry at path
 
         Returns a single dictionary, with exactly the same information as ``ls``
         would with ``detail=True``
+
+        Returns
+        -------
+        dict with keys: name (full path in the FS), size (in bytes), type (file,
+        directory, or something else) and other FS-specific keys.
         """
 
     def isdir(self, path):
         """Is this entry directory-like?"""
+        return self.info(path)['type'] == 'directory'
+
+    def isfile(self, path):
+        """Is this entry file-like?"""
+        return self.info(path)['type'] == 'file'
 
     def cat(self, path):
         """ Get the content of a file """
+        return self.open(path, 'rb').read()
 
     def get(self, rpath, lpath, **kwargs):
         """ Copy file to local
@@ -185,10 +238,14 @@ class AbstractFileSystem(object):
     def copy(self, path1, path2, **kwargs):
         """ Copy within two locations in the filesystem"""
 
+    cp = copy
+
     def mv(self, path1, path2, **kwargs):
         """ Move file from one location to another """
         self.copy(path1, path2, **kwargs)
         self.rm(path1)
+
+    move = mv
 
     def rm(self, path, recursive=False):
         """Delete files.
@@ -202,6 +259,12 @@ class AbstractFileSystem(object):
             also remove the directory
         """
 
+    delete = rm
+
+    def _open(self, path, mode='rb', block_size=None, autocommit=True,
+              **kwargs):
+        pass
+
     def open(self, path, mode='rb', block_size=None, **kwargs):
         """
         Return a file-like object from the filesystem
@@ -211,6 +274,8 @@ class AbstractFileSystem(object):
 
         Parameters
         ----------
+        path: str
+            Target file
         mode: str like 'rb', 'w'
             See builtin ``open()``
         block_size: int
@@ -220,12 +285,22 @@ class AbstractFileSystem(object):
         if 'b' not in mode:
             mode = mode.replace('t', '') + 'b'
             return io.TextIOWrapper(
-                self.open(self, path, mode, block_size, **kwargs))
+                self.open(path, mode, block_size, **kwargs))
+        else:
+            ac = kwargs.pop('autocommit', not self._intrans)
+            f = self._open(path, mode=mode, block_size=block_size,
+                           autocommit=ac, **kwargs)
+            if not ac:
+                self.files.append(f)
+            return f
 
     def touch(self, path, **kwargs):
-        """ Create empty file """
-        with self.open(path, 'wb', **kwargs):
-            pass
+        """ Create empty file, or update timestamp """
+        if not self.exists(path):
+            with self.open(path, 'wb', **kwargs):
+                pass
+        else:
+            raise NotImplementedError  # update timestamp, if possible
 
     def read_block(self, fn, offset, length, delimiter=None):
         """ Read a block of bytes from
